+++ index.ts
@@ -1,1 +1,2 @@
-export {createOpenPromise} from './src/open/promise'
+export type {OpenPromise} from './src/open/promise';
+export {createOpenPromise} from './src/open/promise';