+++ jest.config.ts
@@ -1,7 +1,6 @@
 export default {
 	preset: "ts-jest",
 	testMatch: [
-		"**/__tests__/**/*.[jt]s?(x)",
-		"**/?(*.)+(spec|tests).[tj]s?(x)"
+		"**/?(*.)+(spec|tests).[t]s?(x)"
 	],
 };