+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 -------------------------------
 <a name="start"></a>
 
+### v1.1.0
+
+- Added o-promise `state` property
+- Fixed call `executer` after promise resolving
+
+---
+
 ### v1.0.0
 
 - Base API ;]